requirements.txt
requests>=2.31.0
python-nmap>=0.7.1
pyyaml>=6.0.1
jinja2>=3.1.2
dnspython>=2.4.2
netaddr>=0.9.0
colorama>=0.4.6
tabulate>=0.9.0
cryptography>=41.0.0
